Some small changes.
Worked on playerUI ingame a bit.
Added an abstract property to Spell class for the Sprite imagine icon
which is then overriden by each spell(Fireball for example) where it
returns that spells Icon. Then we can just use like Spell.GetIcon for
the UI, no matter what the spell.